services/command-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
testcontainers>=3.7.1
docker>=6.1.3
redis>=5.0.0
psycopg[binary]>=3.2.1
aiokafka>=0.10.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
